move unit test to store.test
